apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
redis==5.0.8
celery==5.4.0
python-jose==3.3.0
prometheus-fastapi-instrumentator==7.0.0
httpx==0.27.2
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-otlp==1.25.0
pytest==8.3.2
pydantic[email]==2.8.2
passlib[bcrypt]==1.7.4
email-validator==2.2.0
langgraph==0.2.39
langchain-core==0.3.8
jsonschema==4.23.0
reportlab==4.2.2
boto3==1.34.158
PyJWT==2.8.0
#postgresql-client==15.10.0
celery[redis]==5.4.0
